- change the copyright year to 2020 - change levy type enum to uint8 - updated based on review comments by Oleg Oct 27, 2020

+++ plugins/txes/mosaic/src/model/MosaicLevy.h
@@ -12,7 +12,7 @@ namespace catapult { namespace model {
 	constexpr uint64_t MosaicLevyFeeDecimalPlace = 100'000;
 	
 	/// Available mosaic levy rule ids.
-	enum class LevyType : uint16_t {
+	enum class LevyType : uint8_t {
 		/// Default there is no levy
 		None = 0x0,
